# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Contact:
    properties:
      email:
        maxLength: 255
        type: string
      last_name:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      organization_id:
        type: string
      phone:
        maxLength: 255
        type: string
    required:
    - email
    - name
    - organization_id
    - phone
    type: object
  models.DefaultError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  models.Route:
    properties:
      method:
        type: string
      name:
        type: string
      params:
        type: string
      path:
        type: string
    type: object
host: https://newsletter-app-contact-service.herokuapp.com/
info:
  contact:
    email: lasupremaciadelpuntoycoma@gmail.com
    name: API Support
  description: This service manage the contacts of the app.
  title: Contacts microservice for newsletter-app
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Retrieve all routes in this service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Route'
            type: array
      summary: Retrieve all routes
      tags:
      - Routes Public
  /contacts:
    post:
      consumes:
      - application/json
      description: Creates a contact.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DefaultError'
      summary: Creates a contact.
      tags:
      - Contacts
  /contacts/:id:
    delete:
      consumes:
      - application/json
      description: Delete a contact.
      parameters:
      - description: Contact Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.DefaultError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DefaultError'
      summary: Delete a contact.
      tags:
      - Contacts
    get:
      consumes:
      - application/json
      description: Retrieve a contact's data.
      parameters:
      - description: Contact Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.DefaultError'
      summary: Retrieve a contact
      tags:
      - Contacts
    put:
      consumes:
      - application/json
      description: Update a contact.
      parameters:
      - description: Contact Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.DefaultError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.DefaultError'
      summary: Update a contact.
      tags:
      - Contacts
  /contacts/search:
    get:
      consumes:
      - application/json
      description: Search contacts.
      parameters:
      - description: Organization id
        in: query
        name: organization_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Contact'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DefaultError'
      summary: Search contacts.
      tags:
      - Contacts
securityDefinitions:
  "":
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"